+++ actix-web-lab/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ## Unreleased - 2022-xx-xx
-
+- Add `SseSender::send`.
+- Add `SseSender::try_send`.
+- Expose `SseData`.
+- Expose `SseMessage`.
+- Expose `SseSendError`.
+- Deprecate other `SseSender` methods.
 
 ## 0.16.8 - 2022-08-07
 - Add semantic `Sse` (server-sent events) responder with channel-like interface.